kola: update Tang-pinned LUKS tests
- Drop the TPM test since it's already covered by the external host test
  which are inherited from FCOS
- Move to new way of defining LUKS devices
- Dedupe a bunch of things

The test is currently limited to RHCOS but once we have the requisite
patches in dracut and systemd in FCOS, we should be able to target it
too:

https://github.com/dracutdevs/dracut/pull/931
https://github.com/dracutdevs/dracut/pull/968
https://github.com/systemd/systemd/pull/17467
